feat: add VirtualSvrollbar
Contributes a version of Svrollbar which doesn't setup viewport or
content listeners to determine the height and progress of the scrollbar
but instead accepts props to provide those values.

This is useful in combination with virtual lists as the height calculation
is already performed there so it doesn't need to be performed again in
the scrollbar.